Todos int agora sao long. Problema na matriz A em Contagem.c depois que passa em ordena

+++ 4_BioInformatica/bio.c
@@ -10,7 +10,7 @@
 
 #include "bio.h"
 
-void ContagemLeituras (FILE* arquivo_genoma, FILE* arquivo_pos_genes, FILE* arquivo_fragmentos, FILE* arquivo_pos_fragmentos, int n_genes, int n_fragmentos, FILE* arquivo_saida) {
+void ContagemLeituras (FILE* arquivo_genoma, FILE* arquivo_pos_genes, FILE* arquivo_fragmentos, FILE* arquivo_pos_fragmentos, long n_genes, long n_fragmentos, FILE* arquivo_saida) {
   CtrlF(arquivo_genoma, arquivo_fragmentos, arquivo_pos_fragmentos);
   fclose(arquivo_pos_fragmentos);
   FILE* arquivo_pos_fragmentos2 = fopen("saida.txt", "r");
