+++ 4_BioInformatica/bio.h
@@ -10,4 +10,4 @@
 #include "contagem.h"
 #include "ctrlf.h"
 
-void ContagemLeituras (FILE* arquivo_genoma, FILE* arquivo_fragmentos, FILE* arquivo_pos_genes, FILE* arquivo_pos_fragmentos, int n_genes, int n_fragmentos, FILE* arquivo_saida);
+void ContagemLeituras (FILE* arquivo_genoma, FILE* arquivo_fragmentos, FILE* arquivo_pos_genes, FILE* arquivo_pos_fragmentos, long n_genes, long n_fragmentos, FILE* arquivo_saida);
